refactor: move sidebar help support
